add GraphQLSchema
