+++ packages/fuel-indexer-macros/src/lib.rs
@@ -1,7 +1,7 @@
 extern crate lazy_static;
 extern crate proc_macro;
 
-pub(crate) mod constant;
+pub(crate) mod constants;
 pub(crate) mod helpers;
 pub(crate) mod indexer;
 pub(crate) mod native;
